Remove inheritance for InferIsNotNull
Inherit directly from Rule<P, Q>, do not override skipExpression but
directly set it to skip Coalesce only.
